+++ src/PlayerEngagement.Infrastructure/Policies/Mappers/PolicySegmentOverrideMapper.cs
@@ -5,10 +5,10 @@ using PlayerEngagement.Infrastructure.Persistence.DTOs.XpPolicyDTOs;
 namespace PlayerEngagement.Infrastructure.Policies.Mappers;
 
 internal static class PolicySegmentOverrideMapper {
-    internal static IReadOnlyDictionary<string, int> ToDictionary(IEnumerable<PolicySegmentOverrideDTO> overrides) {
+    internal static IReadOnlyDictionary<string, long> ToDictionary(IEnumerable<PolicySegmentOverrideDTO> overrides) {
         ArgumentNullException.ThrowIfNull(overrides);
 
-        Dictionary<string, int> result = new(StringComparer.OrdinalIgnoreCase);
+        Dictionary<string, long> result = new(StringComparer.OrdinalIgnoreCase);
         foreach (PolicySegmentOverrideDTO dto in overrides) {
             if (dto.IsEmpty)
                 continue;
